LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/get-persistendeviceid-from-pointer-event.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_equals: deviceId is 0 expected (number) 0 but got (undefined) undefined
CONSOLE MESSAGE: Error: assert_equals: deviceId is 0 expected (number) 0 but got (undefined) undefined

Harness Error (FAIL), message = Error: assert_equals: deviceId is 0 expected (number) 0 but got (undefined) undefined

PASS PointerEvent.persistentDeviceId
